Update from netsec discussion 30 Mar 2021
Add delegated third party language to 5.5.2. Modify 5.5.1 language slightly.
Clarify expectations surrounding the undefined "OCSP entries" term; that is, the CA needs to be logging each time they use a key to sign something, so a log entry for an "OCSP entry" might contain the timestamp+action, an identifier of the certificate whose status is being signed, and the state of revocation being signed, but need not necessarily include the full OCSP response itself in such an audit log.